Move mockForkHistory into separate file
