+++ packages/graphql_server/src/utils.ts
@@ -79,102 +79,3 @@ export const parseGitHubUrl = (url: string) => {
 
   return null
 }
-
-export const mockForkHistory = [
-  {
-    date: '2023/6/13 15:11:54',
-    count: 30
-  },
-  {
-    date: '2023/6/13 15:41:52',
-    count: 90
-  },
-  {
-    date: '2023/6/13 16:5:18',
-    count: 120
-  },
-  {
-    date: '2023/6/13 16:30:57',
-    count: 150
-  },
-  {
-    date: '2023/6/13 16:57:44',
-    count: 180
-  },
-  {
-    date: '2023/6/13 17:36:43',
-    count: 210
-  },
-  {
-    date: '2023/6/13 18:14:46',
-    count: 240
-  },
-  {
-    date: '2023/6/13 19:57:5',
-    count: 300
-  },
-  {
-    date: '2023/6/13 21:50:46',
-    count: 330
-  },
-  {
-    date: '2023/6/14 0:28:33',
-    count: 360
-  },
-  {
-    date: '2023/6/14 3:9:57',
-    count: 390
-  },
-  {
-    date: '2023/6/14 4:44:44',
-    count: 420
-  },
-  {
-    date: '2023/6/14 6:42:5',
-    count: 450
-  },
-  {
-    date: '2023/6/14 11:35:50',
-    count: 510
-  },
-  {
-    date: '2023/6/14 15:9:25',
-    count: 540
-  },
-  {
-    date: '2023/6/14 21:18:18',
-    count: 570
-  },
-  {
-    date: '2023/6/15 4:22:36',
-    count: 600
-  },
-  {
-    date: '2023/6/15 14:10:51',
-    count: 630
-  },
-  {
-    date: '2023/6/15 17:16:37',
-    count: 660
-  },
-  {
-    date: '2023/6/16 4:32:46',
-    count: 720
-  },
-  {
-    date: '2023/6/16 8:1:2',
-    count: 750
-  },
-  {
-    date: '2023/6/16 11:23:30',
-    count: 780
-  },
-  {
-    date: '2023/6/16 16:20:50',
-    count: 810
-  },
-  {
-    date: '2023/6/17 1:39:55',
-    count: 843
-  }
-]
